Moved things into their own namespace for the Http namespace
